@@ -1,20 +1,19 @@
 package pro.gravit.simplecabinet.web.service.storage;
 
+import jakarta.annotation.Priority;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 import pro.gravit.simplecabinet.web.configuration.properties.FileStorageConfig;
 
-import javax.annotation.Priority;
 import java.io.*;
-import java.net.MalformedURLException;
-import java.net.URL;
-import java.net.URLEncoder;
+import java.net.*;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Path;
 
 @Service
 @Priority(value = 1)
-public class FileStorageService implements StorageService {
+public class
+FileStorageService implements StorageService {
 
     @Autowired
     private FileStorageConfig config;
@@ -44,8 +43,8 @@ public class FileStorageService implements StorageService {
     @Override
     public URL getUrl(String name) {
         try {
-            return new URL(config.getRemoteUrl().concat(URLEncoder.encode(name, StandardCharsets.UTF_8)));
-        } catch (MalformedURLException e) {
+            return new URI(config.getRemoteUrl().concat(URLEncoder.encode(name, StandardCharsets.UTF_8))).toURL();
+        } catch (MalformedURLException | URISyntaxException e) {
             throw new RuntimeException(e);
         }
     }
