@@ -3,7 +3,6 @@ package pro.gravit.simplecabinet.web.configuration.jwt;
 import io.jsonwebtoken.Claims;
 import io.jsonwebtoken.JwtBuilder;
 import io.jsonwebtoken.Jwts;
-import io.jsonwebtoken.SignatureAlgorithm;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import pro.gravit.simplecabinet.web.model.user.User;
@@ -29,7 +28,7 @@ public class JwtProvider {
     public GeneratedJWTToken generateToken(UserSession session) {
         LocalDateTime dateTime = LocalDateTime.now().plusDays(15);
         var token = makeBasic(session)
-                .setExpiration(Date.from(dateTime.toInstant(ZoneOffset.UTC)))
+                .expiration(Date.from(dateTime.toInstant(ZoneOffset.UTC)))
                 .compact();
         return new GeneratedJWTToken(token, dateTime);
     }
@@ -45,19 +44,19 @@ public class JwtProvider {
         User user = session.getUser();
         var roles = userDetailsService.collectUserRoles(user);
         return Jwts.builder()
-                .setSubject(user.getUsername())
-                .setIssuer("SimpleCabinet")
+                .subject(user.getUsername())
+                .issuer("SimpleCabinet")
                 .claim("roles", roles)
                 .claim("id", user.getId())
                 .claim("sessionId", session.getId())
                 .claim("client", session.getClient())
-                .signWith(keyManagementService.getPrivateKey(), SignatureAlgorithm.ES256);
+                .signWith(keyManagementService.getPrivateKey(), Jwts.SIG.ES256);
     }
 
     @SuppressWarnings("unchecked")
     public UserDetailsService.CabinetUserDetails getDetailsFromToken(String token) {
         Claims claims = parserProvider.makeParser()
-                .parseClaimsJws(token).getBody();
+                .parseSignedClaims(token).getPayload();
         List<String> roles = claims.get("roles", List.class);
         String client = claims.get("client", String.class);
         long userId = claims.get("id", Long.class);
@@ -67,14 +66,14 @@ public class JwtProvider {
 
     public boolean validateToken(String token) {
         try {
-            parserProvider.makeParser().parseClaimsJws(token);
+            parserProvider.makeParser().parseSignedClaims(token);
             return true;
         } catch (Exception ignored) {
         }
         return false;
     }
 
-    public static record GeneratedJWTToken(String token, LocalDateTime endTime) {
+    public record GeneratedJWTToken(String token, LocalDateTime endTime) {
         public long getExpire() {
             return Duration.between(LocalDateTime.now(), endTime).toMillis();
         }
