@@ -1,5 +1,6 @@
 package pro.gravit.simplecabinet.web.controller.payment;
 
+import jakarta.servlet.http.HttpServletRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.web.util.matcher.IpAddressMatcher;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -7,8 +8,6 @@ import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 import pro.gravit.simplecabinet.web.service.payment.YooPaymentService;
-
-import javax.servlet.http.HttpServletRequest;
 
 @RestController
 @RequestMapping("/webhooks/yoo")
