@@ -1,8 +1,7 @@
 package pro.gravit.simplecabinet.web.model;
 
+import jakarta.persistence.*;
 import pro.gravit.simplecabinet.web.model.user.User;
-
-import javax.persistence.*;
 
 @Entity(name = "NewsComment")
 @Table(name = "news_comments")
