@@ -1,10 +1,9 @@
 package pro.gravit.simplecabinet.web.service.mail;
 
+import jakarta.annotation.Priority;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
-
-import javax.annotation.Priority;
 
 @Service
 @Priority(value = 1)
