@@ -1,11 +1,10 @@
 package pro.gravit.simplecabinet.web.service.mail;
 
+import jakarta.annotation.Priority;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.autoconfigure.condition.ConditionalOnProperty;
 import org.springframework.mail.javamail.JavaMailSender;
 import org.springframework.stereotype.Service;
-
-import javax.annotation.Priority;
 
 @Service
 @Priority(value = 0)
