@@ -7,6 +7,7 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 import pro.gravit.simplecabinet.web.configuration.properties.FreekassaPaymentConfig;
 import pro.gravit.simplecabinet.web.exception.PaymentException;
 import pro.gravit.simplecabinet.web.model.shop.Payment;
@@ -14,7 +15,6 @@ import pro.gravit.simplecabinet.web.model.user.User;
 import pro.gravit.simplecabinet.web.service.shop.PaymentService;
 import pro.gravit.simplecabinet.web.utils.SecurityUtils;
 
-import javax.transaction.Transactional;
 import java.io.IOException;
 import java.net.URI;
 import java.net.http.HttpClient;
